customer edit yapıldı

+++ eLonca.Common/DTOs/SaleItemDto.cs
@@ -9,7 +9,7 @@
         public decimal? UnitPrice { get; set; }
         public decimal? TotalAmount { get; set; }
         public decimal? Discount { get; set; }
-        public decimal? CustomerDiscount { get; set; }
+        public double? CustomerDiscount { get; set; }
         public decimal? TotalPrice { get; set; }
         public string ProductName { get; set; }
         public string ProductCode { get; set; }
